--- api/src/opentrons/drivers/asyncio/tempdeck/__init__.py
@@ -1,10 +1,0 @@
-from .driver import TempDeckDriver
-from .abstract import AbstractTempDeckDriver
-from .simulator import SimulatingDriver
-
-
-__all__ = [
-    "TempDeckDriver",
-    "AbstractTempDeckDriver",
-    "SimulatingDriver",
-]
